+++ Engine/Poo.h
@@ -6,15 +6,15 @@
 class Poo
 {
 public:
-	void Init( int in_x,int in_y,int in_vx,int in_vy );
+	void Init( float in_x, float in_y, float in_vx, float in_vy );
 	void Update();
 	bool TestCollision( const Dude& dude ) const;
 	void Draw( Graphics& gfx ) const;
 private:
-	int x;
-	int y;
-	int vx;
-	int vy;
+	float x;
+	float y;
+	float vx;
+	float vy;
 	static constexpr int width = 24;
 	static constexpr int height = 24;
 	bool initialized = false;
